Separate the note property change handler from the notes change handler. Also, fix the test case broken.
