+++ osu.Game.Rulesets.Karaoke.Tests/Editor/ChangeHandlers/Notes/NotesChangeHandlerTest.cs
@@ -83,69 +83,6 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.ChangeHandlers.Notes
         }
 
         [Test]
-        public void TestChangeText()
-        {
-            PrepareHitObject(new Note
-            {
-                Text = "カラオケ",
-            });
-
-            TriggerHandlerChanged(c => c.ChangeText("からおけ"));
-
-            AssertSelectedHitObject(h =>
-            {
-                Assert.AreEqual("からおけ", h.Text);
-            });
-        }
-
-        [Test]
-        public void TestChangeRubyText()
-        {
-            PrepareHitObject(new Note
-            {
-                RubyText = "からおけ",
-            });
-
-            TriggerHandlerChanged(c => c.ChangeRubyText("カラオケ"));
-
-            AssertSelectedHitObject(h =>
-            {
-                Assert.AreEqual("カラオケ", h.RubyText);
-            });
-        }
-
-        [Test]
-        public void TestChangeDisplayStateToVisible()
-        {
-            PrepareHitObject(new Note());
-
-            TriggerHandlerChanged(c => c.ChangeDisplayState(true));
-
-            AssertSelectedHitObject(h =>
-            {
-                Assert.IsTrue(h.Display);
-            });
-        }
-
-        [Test]
-        public void TestChangeDisplayStateToNonVisible()
-        {
-            PrepareHitObject(new Note
-            {
-                Display = true,
-                Tone = new Tone(3)
-            });
-
-            TriggerHandlerChanged(c => c.ChangeDisplayState(false));
-
-            AssertSelectedHitObject(h =>
-            {
-                Assert.IsFalse(h.Display);
-                Assert.AreEqual(new Tone(), h.Tone);
-            });
-        }
-
-        [Test]
         public void TestClear()
         {
             var lyric = new Lyric();
