+++ osu.Game.Rulesets.Karaoke/Edit/KaraokeSelectionHandler.cs
@@ -44,6 +44,9 @@ namespace osu.Game.Rulesets.Karaoke.Edit
         private INotesChangeHandler notesChangeHandler { get; set; }
 
         [Resolved]
+        private INotePropertyChangeHandler notePropertyChangeHandler { get; set; }
+
+        [Resolved]
         private ILyricSingerChangeHandler lyricSingerChangeHandler { get; set; }
 
         protected ScrollingNotePlayfield NotePlayfield => ((KaraokeHitObjectComposer)composer).Playfield.NotePlayfield;
@@ -85,7 +88,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit
             return new OsuMenuItem($"{displayText} {selectedObject.Count} notes.", display ? MenuItemType.Destructive : MenuItemType.Standard,
                 () =>
                 {
-                    notesChangeHandler.ChangeDisplayState(!display);
+                    notePropertyChangeHandler.ChangeDisplayState(!display);
                 });
         }
