+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Extends/Notes/NoteEditPropertySection.cs
@@ -88,7 +88,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
         private class LabelledNoteTextTextBox : LabelledObjectFieldTextBox<Note>
         {
             [Resolved]
-            private INotesChangeHandler notesChangeHandler { get; set; }
+            private INotePropertyChangeHandler notePropertyChangeHandler { get; set; }
 
             public LabelledNoteTextTextBox(Note item)
                 : base(item)
@@ -99,7 +99,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
                 => note.Text;
 
             protected override void ApplyValue(Note note, string value)
-                => notesChangeHandler.ChangeText(value);
+                => notePropertyChangeHandler.ChangeText(value);
 
             [BackgroundDependencyLoader]
             private void load(IEditNoteModeState editNoteModeState)
@@ -111,7 +111,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
         private class LabelledNoteRubyTextTextBox : LabelledObjectFieldTextBox<Note>
         {
             [Resolved]
-            private INotesChangeHandler notesChangeHandler { get; set; }
+            private INotePropertyChangeHandler notePropertyChangeHandler { get; set; }
 
             public LabelledNoteRubyTextTextBox(Note item)
                 : base(item)
@@ -122,7 +122,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
                 => note.RubyText;
 
             protected override void ApplyValue(Note note, string value)
-                => notesChangeHandler.ChangeRubyText(value);
+                => notePropertyChangeHandler.ChangeRubyText(value);
 
             [BackgroundDependencyLoader]
             private void load(IEditNoteModeState editNoteModeState)
@@ -134,7 +134,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
         private class LabelledNoteDisplaySwitchButton : LabelledObjectFieldSwitchButton<Note>
         {
             [Resolved]
-            private INotesChangeHandler notesChangeHandler { get; set; }
+            private INotePropertyChangeHandler notePropertyChangeHandler { get; set; }
 
             public LabelledNoteDisplaySwitchButton(Note item)
                 : base(item)
@@ -145,7 +145,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
                 => note.Display;
 
             protected override void ApplyValue(Note note, bool value)
-                => notesChangeHandler.ChangeDisplayState(value);
+                => notePropertyChangeHandler.ChangeDisplayState(value);
 
             [BackgroundDependencyLoader]
             private void load(IEditNoteModeState editNoteModeState)
